move dcs to plot
