+++ nlmod/__init__.py
@@ -8,6 +8,6 @@ import os
 
 NLMOD_DATADIR = os.path.join(os.path.dirname(__file__), "data")
 
-from . import dcs, dims, gis, gwf, gwt, modpath, plot, read, sim, util
+from . import dims, gis, gwf, gwt, modpath, plot, read, sim, util
 from .dims import base, get_ds, grid, layers, resample, time, to_model_ds
 from .version import __version__, show_versions
