+++ nlmod/plot/plot.py
@@ -11,7 +11,7 @@ from matplotlib.colors import ListedColormap, Normalize
 from matplotlib.patches import Patch
 from mpl_toolkits.axes_grid1 import make_axes_locatable
 
-from ..dcs import DatasetCrossSection
+from .dcs import DatasetCrossSection
 from ..dims.grid import modelgrid_from_ds
 from ..dims.resample import get_affine_mod_to_world, get_extent
 from ..read import geotop, rws
@@ -120,7 +120,8 @@ def facet_plot(
         iax = axes.ravel()[ilay]
         mp = fp.plot.PlotMapView(model=gwf, layer=ilay, ax=iax)
         # mp.plot_grid()
-        qm = mp.plot_array(plot_arr[ilay].values, cmap="viridis", vmin=vmin, vmax=vmax)
+        qm = mp.plot_array(plot_arr[ilay].values,
+                           cmap="viridis", vmin=vmin, vmax=vmax)
         # qm = mp.plot_array(hf[-1], cmap="viridis", vmin=-0.1, vmax=0.1)
         # mp.plot_ibound()
         # plt.colorbar(qm)
@@ -403,7 +404,7 @@ def map_array(
 
     if save:
         f.savefig(fname, bbox_inches="tight", dpi=150)
-    
+
     return ax
